Sujith :) -> 1. Implemented translate template 2. Bunch of refactoring 3. Fixed on click bug not unsubscribing

+++ Assets/_Studio/_Runtime/Scripts/Components/RotateComponent.cs
@@ -27,7 +27,7 @@ namespace Terra.Studio
         public BroadcastAt broadcastAt;
         public float speed;
         public float rotateBy;
-        public int pauseFor;
+        public float pauseFor;
         public int repeatFor;
         public bool canPlaySFX;
         public string sfxName;
